+++ lib/bones_api_console.dart
@@ -1,6 +1,6 @@
-/// Bones_API Server Library.
+/// Bones_API Console Library.
 library bones_api_server;
 
 export 'package:bones_api/bones_api.dart';
 
-export 'src/bones_api_server.dart';
+export 'src/bones_api_console.dart';