+++ lib/bones_api_dart_spawner.dart
@@ -2,5 +2,4 @@
 library bones_api_server;
 
 export 'package:bones_api/bones_api.dart';
-
-export 'src/bones_api_server.dart';
+export 'package:dart_spawner/dart_spawner.dart';